HW6Backend/pyproject.toml
[project]
name = "hw6backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Nurlibaj",email = "nurlibaj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "jinja2 (>=3.1.5,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
